requirements.txt
discord.py==2.3.2
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.0
anthropic==0.34.0
apscheduler==3.10.4
pyyaml==6.0.2
python-dotenv==1.0.1
PyNaCl==1.5.0
